pyproject.toml
[project]
name = "Snake"
version = "0.1.0"
description = "A Snake game"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "pygame>=2.6.1",
]
